refactor onboarding page
