+++ src/pages/landing/types.ts
