+++ src/pages/onboarding/components/OnboardingFlow.tsx
@@ -1,31 +1,32 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { initializePaddle, type Paddle } from '@paddle/paddle-js';
 
-import Button from '../../../components/Button';
-import Header from '../../../components/Header';
-import { CircleAlert, CircleCheck } from 'lucide-react';
-
-import { useAuth } from '../../../hooks/useAuth';
 import { getAuthService } from '../../../utils/AuthService';
+import { useAuth } from '../../../hooks/useAuth';
 
-import { PUBLIC_PADDLE_CLIENT_TOKEN } from 'astro:env/client';
-import { PUBLIC_PADDLE_PRICE_ID } from 'astro:env/client';
+import {
+  PUBLIC_CF_ENV,
+  PUBLIC_PADDLE_CLIENT_TOKEN,
+  PUBLIC_PADDLE_PRICE_ID,
+} from 'astro:env/client';
+
+import Header from '../../../components/Header';
+import OnboardingInitialView from './OnboardingInitialView';
+import OnboardingSuccessView from './OnboardingSuccessView';
+import OnboardingErrorView from './OnboardingErrorView';
 
 const OnboardingFlow: React.FC = () => {
   const authService = getAuthService()
-  const { isLoggedIn, user, login } = useAuth();
 
+  const { isLoggedIn, user, login} = useAuth();
   const [currentView, setCurrentView] = useState<'initial' | 'success' | 'error'>('initial');
-  const [discountCode, setDiscountCode] = useState('');
 
+  const [discountCode, setDiscountCode] = useState('');
   const [userFriendlyError, setUserFriendlyError] = useState('We encountered an unexpected issue.');
   const [detailedErrorMessage, setDetailedErrorMessage] = useState<string | null>(null);
-  const [showDetailedError, setShowDetailedError] = useState(false);
 
   const paddle = useRef<Paddle | undefined>(undefined);
-  const env = import.meta.env;
 
-  // --- METHODS ---
   const handleError = (userMessage: string, errorObj: any = null) => {
     console.error('User-facing error triggered:', userMessage, errorObj || '');
     setUserFriendlyError(userMessage || 'An unexpected error occurred.');
@@ -41,7 +42,6 @@ const OnboardingFlow: React.FC = () => {
     } else {
       setDetailedErrorMessage(null);
     }
-    setShowDetailedError(false);
     setCurrentView('error');
   };
 
@@ -49,8 +49,7 @@ const OnboardingFlow: React.FC = () => {
     try {
       if (!paddle.current) {
         const environment =
-          (env.PUBLIC_PADDLE_ENVIRONMENT as 'sandbox' | 'production') ||
-          'sandbox';
+          (PUBLIC_CF_ENV as 'sandbox' | 'production') || 'sandbox';
         const clientToken = PUBLIC_PADDLE_CLIENT_TOKEN;
         if (!clientToken)
           throw new Error('Paddle client token is not configured');
@@ -78,9 +77,7 @@ const OnboardingFlow: React.FC = () => {
 
   const openPaddleCheckout = async () => {
     try {
-      if (!paddle.current) {
-        await setupPaddleCheckout();
-      }
+      if (!paddle.current) await setupPaddleCheckout();
       await paddle.current?.Checkout.open({
         items: [{ priceId: PUBLIC_PADDLE_PRICE_ID, quantity: 1 }],
         ...(discountCode && { discountCode }),
@@ -93,8 +90,6 @@ const OnboardingFlow: React.FC = () => {
   };
 
   // --- LIFECYCLE ---
-
-  // Effect to handle view changes based on authentication state
   useEffect(() => {
     if (isLoggedIn) {
       setCurrentView('success');
@@ -106,18 +101,33 @@ const OnboardingFlow: React.FC = () => {
     }
   }, [isLoggedIn]);
 
-  // Effect to grab discount code from URL on initial load
   useEffect(() => {
     const urlParams = new URLSearchParams(window.location.search);
     const discountFromQuery = urlParams.get('discount');
-    if (discountFromQuery) {
-      setDiscountCode(discountFromQuery);
-    }
+    if (discountFromQuery) setDiscountCode(discountFromQuery);
   }, []);
 
-  const isTrialing =
-    user?.profile.context?.user?.subscription_status !== 'active';
-  const firstName = user?.profile.name?.split(' ')[0] ?? 'there';
+  const renderCurrentView = () => {
+    switch (currentView) {
+      case 'success':
+        return (
+          user && (
+            <OnboardingSuccessView user={user} onOpenCheckout={openPaddleCheckout} />
+          )
+        );
+      case 'error':
+        return (
+          <OnboardingErrorView
+            userFriendlyError={userFriendlyError}
+            detailedErrorMessage={detailedErrorMessage}
+            onTryAgain={() => setCurrentView('initial')}
+          />
+        );
+      case 'initial':
+      default:
+        return <OnboardingInitialView onLogin={login} onSignup={() => authService.signup()} />;
+    }
+  };
 
   return (
     <>
@@ -125,148 +135,7 @@ const OnboardingFlow: React.FC = () => {
 
       <div className='container mx-auto px-4 py-12'>
         <div className='max-w-lg mx-auto bg-white rounded-2xl shadow-lg border border-gray-200 p-8 md:p-12'>
-          {/* --- Initial View --- */}
-          {currentView === 'initial' && (
-            <div className='text-center'>
-              <h1 className='text-4xl font-bold text-gray-900 mb-4'>
-                Get Started
-              </h1>
-              <p className='text-lg text-gray-600 mb-8'>
-                Create an account or log in to continue.
-              </p>
-              <div className='space-y-4'>
-                <Button
-                  onClick={() => authService.signup()}
-                  variant='primary'
-                  className='w-full'
-                >
-                  Create a New Account
-                </Button>
-                <Button
-                  onClick={() => login()}
-                  variant='secondary'
-                  className='w-full'
-                >
-                  Log In
-                </Button>
-              </div>
-              <p className='text-sm text-gray-500 mt-6'>
-                Need help? Contact us at{' '}
-                <a
-                  href='mailto:support@sonacove.com'
-                  className='font-medium text-primary-600 hover:underline'
-                >
-                  support@sonacove.com
-                </a>
-              </p>
-            </div>
-          )}
-
-          {/* --- Success View --- */}
-          {currentView === 'success' && user && (
-            <div className='text-center'>
-              <CircleCheck className='w-16 h-16 mx-auto text-green-500' />
-              <h1 className='text-4xl font-bold text-gray-900 mt-6 mb-2'>
-                Welcome aboard, {firstName}!
-              </h1>
-
-              {isTrialing ? (
-                <>
-                  <p className='text-lg text-gray-600'>
-                    You're now on our unlimited free trial.
-                  </p>
-                  <div className='bg-gray-50 rounded-xl p-6 my-8 text-left space-y-2 border'>
-                    <p className='font-semibold text-gray-800'>Your Account:</p>
-                    <p className='text-gray-600'>
-                      <strong>Email:</strong> {user.profile.email}
-                    </p>
-                    <p className='text-gray-600'>
-                      <strong>Status:</strong> Free Trial
-                    </p>
-                  </div>
-                  <div className='space-y-4 flex flex-col'>
-                    <Button
-                      onClick={openPaddleCheckout}
-                      variant='primary'
-                      className='w-full'
-                    >
-                      Subscribe Now for Full Access
-                    </Button>
-                    <a href='/meet'>
-                      <Button variant='secondary' className='w-full'>
-                        Continue with Trial
-                      </Button>
-                    </a>
-                  </div>
-                </>
-              ) : (
-                <>
-                  <p className='text-lg text-gray-600'>
-                    Your account is fully activated.
-                  </p>
-                  <div className='bg-gray-50 rounded-xl p-6 my-8 text-left space-y-2 border'>
-                    <p className='font-semibold text-gray-800'>Your Account:</p>
-                    <p className='text-gray-600'>
-                      <strong>Email:</strong> {user.profile.email}
-                    </p>
-                    <p className='text-gray-600'>
-                      <strong>Status:</strong> Active
-                    </p>
-                  </div>
-                  <a href='/meet'>
-                    <Button variant='primary' className='w-full'>
-                      Go to Sonacove Meets
-                    </Button>
-                  </a>
-                </>
-              )}
-            </div>
-          )}
-
-          {/* --- Error View --- */}
-          {currentView === 'error' && (
-            <div className='text-center'>
-              <CircleAlert className='w-16 h-16 mx-auto text-red-500' />
-              <h1 className='text-4xl font-bold text-gray-900 mt-6 mb-2'>
-                Something went wrong
-              </h1>
-              <p className='text-lg text-gray-600 mb-6'>{userFriendlyError}</p>
-
-              {detailedErrorMessage && (
-                <div className='mb-6 text-sm text-left bg-red-50 p-4 rounded-lg border border-red-200'>
-                  <button
-                    onClick={() => setShowDetailedError(!showDetailedError)}
-                    className='font-semibold text-primary-600 hover:underline mb-2'
-                  >
-                    {showDetailedError
-                      ? 'Hide Technical Details'
-                      : 'Show Technical Details'}
-                  </button>
-                  {showDetailedError && (
-                    <pre className='mt-2 text-xs text-gray-600 whitespace-pre-wrap bg-white p-2 rounded'>
-                      {detailedErrorMessage}
-                    </pre>
-                  )}
-                </div>
-              )}
-
-              <div className='mt-8 space-y-4'>
-                  <Button
-                    onClick={() => authService.signup()}
-                    variant='primary'
-                    className='w-full'
-                  >
-                    Try Again
-                  </Button>
-                <a
-                  href='mailto:support@sonacove.com'
-                  className='block text-center text-primary-600 font-medium hover:underline'
-                >
-                  Contact Support
-                </a>
-              </div>
-            </div>
-          )}
+          {renderCurrentView()}
         </div>
       </div>
     </>
